+++ src/Ydb.Sdk/src/Ado/PoolManager.cs
@@ -6,7 +6,9 @@ namespace Ydb.Sdk.Ado;
 internal static class PoolManager
 {
     private static readonly SemaphoreSlim SemaphoreSlim = new(1); // async mutex
-    private static readonly ConcurrentDictionary<string, ISessionSource> Pools = new();
+
+    internal static readonly ConcurrentDictionary<string, IDriver> Drivers = new();
+    internal static readonly ConcurrentDictionary<string, ISessionSource> Pools = new();
 
     internal static async Task<ISession> GetSession(
         YdbConnectionStringBuilder settings,
@@ -18,18 +20,23 @@ internal static class PoolManager
             return await sessionPool.OpenSession(cancellationToken);
         }
 
+        await SemaphoreSlim.WaitAsync(cancellationToken);
+
         try
         {
-            await SemaphoreSlim.WaitAsync(cancellationToken);
-
             if (Pools.TryGetValue(settings.ConnectionString, out var pool))
             {
                 return await pool.OpenSession(cancellationToken);
             }
 
-            var newSessionPool = new PoolingSessionSource<PoolingSession>(
-                await PoolingSessionFactory.Create(settings), settings
-            );
+            var driver = Drivers.TryGetValue(settings.GrpcConnectionString, out var cacheDriver) &&
+                         !cacheDriver.IsDisposed
+                ? cacheDriver
+                : Drivers[settings.GrpcConnectionString] = await settings.BuildDriver();
+            driver.RegisterOwner();
+
+            var factory = new PoolingSessionFactory(driver, settings);
+            var newSessionPool = new PoolingSessionSource<PoolingSession>(factory, settings);
 
             Pools[settings.ConnectionString] = newSessionPool;
 
@@ -43,12 +50,11 @@ internal static class PoolManager
 
     internal static async Task ClearPool(string connectionString)
     {
-        if (Pools.Remove(connectionString, out var sessionPool))
+        if (Pools.TryRemove(connectionString, out var sessionPool))
         {
+            await SemaphoreSlim.WaitAsync();
             try
             {
-                await SemaphoreSlim.WaitAsync();
-
                 await sessionPool.DisposeAsync();
             }
             finally