+++ src/Ydb.Sdk/src/Ado/YdbConnectionStringBuilder.cs
@@ -3,6 +3,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Security.Cryptography.X509Certificates;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
+using Ydb.Sdk.Ado.RetryPolicy;
 using Ydb.Sdk.Auth;
 using Ydb.Sdk.Transport;
 
@@ -314,11 +315,15 @@ public sealed class YdbConnectionStringBuilder : DbConnectionStringBuilder
 
     private int _createSessionTimeout;
 
-    public ILoggerFactory? LoggerFactory { get; init; }
+    public ILoggerFactory LoggerFactory { get; init; } = NullLoggerFactory.Instance;
 
     public ICredentialsProvider? CredentialsProvider { get; init; }
 
     public X509Certificate2Collection? ServerCertificates { get; init; }
+
+    public IRetryPolicy RetryPolicy { get; init; } = YdbRetryPolicy.Default;
+
+    internal YdbRetryPolicyExecutor YdbRetryPolicyExecutor => new(RetryPolicy);
 
     private void SaveValue(string propertyName, object? value)
     {
@@ -358,6 +363,12 @@ public sealed class YdbConnectionStringBuilder : DbConnectionStringBuilder
 
     private string Endpoint => $"{(UseTls ? "grpcs" : "grpc")}://{Host}:{Port}";
 
+    internal string GrpcConnectionString =>
+        $"UseTls={UseTls};Host={Host};Port={Port};Database={Database};User={User};Password={Password};" +
+        $"ConnectTimeout={ConnectTimeout};KeepAlivePingDelay={KeepAlivePingDelay};KeepAlivePingTimeout={KeepAlivePingTimeout};" +
+        $"EnableMultipleHttp2Connections={EnableMultipleHttp2Connections};MaxSendMessageSize={MaxSendMessageSize};" +
+        $"MaxReceiveMessageSize={MaxReceiveMessageSize};DisableDiscovery={DisableDiscovery}";
+
     internal async Task<IDriver> BuildDriver()
     {
         var cert = RootCertificate != null ? X509Certificate.CreateFromCertFile(RootCertificate) : null;
@@ -384,11 +395,10 @@ public sealed class YdbConnectionStringBuilder : DbConnectionStringBuilder
             MaxSendMessageSize = MaxSendMessageSize,
             MaxReceiveMessageSize = MaxReceiveMessageSize
         };
-        var loggerFactory = LoggerFactory ?? NullLoggerFactory.Instance;
 
         return DisableDiscovery
-            ? new DirectGrpcChannelDriver(driverConfig, loggerFactory)
-            : await Driver.CreateInitialized(driverConfig, loggerFactory);
+            ? new DirectGrpcChannelDriver(driverConfig, LoggerFactory)
+            : await Driver.CreateInitialized(driverConfig, LoggerFactory);
     }
 
     public override void Clear()