+++ src/Ydb.Sdk/test/Ydb.Sdk.Ado.Tests/Utils/DisableParallelization.cs
@@ -1,0 +1,6 @@
+using Xunit;
+
+namespace Ydb.Sdk.Ado.Tests.Utils;
+
+[CollectionDefinition("DisableParallelization", DisableParallelization = true)]
+public class DisableParallelization;